Reformat

+++ tests/cupy_tests/core_tests/fusion_tests/test_misc.py
@@ -74,8 +74,8 @@ class TestFusionTuple(FusionTestBase):
     @fusion_utils.check_fusion(generate_inputs_args=(3,))
     def test_multiple_outputdifferent_type_same_ufunc(self, xp, dtype):
         def func(x, y, z):
-            x = x.astype('int32')
-            y = x.astype('float32')
+            x = x.astype("int32")
+            y = x.astype("float32")
             return x + y, y + z, z + x
 
         return func
@@ -98,7 +98,8 @@ class TestFusionTuple(FusionTestBase):
 
     @unittest.skipUnless(
         fusion_utils.can_use_grid_synchronization(),
-        'Requires CUDA grid synchronization')
+        "Requires CUDA grid synchronization",
+    )
     @testing.for_all_dtypes(no_bool=True, no_complex=True)
     @fusion_utils.check_fusion(generate_inputs_args=(2,))
     def test_various_shape(self, xp, dtype):
@@ -109,6 +110,7 @@ class TestFusionTuple(FusionTestBase):
             c = xp.sum(x, axis=0)
             y += c
             return c, b, x, y, a
+
         return func
 
 
@@ -119,6 +121,7 @@ class TestReturnNone(FusionTestBase):
     def test_pass(self, xp, dtype):
         def func():
             pass
+
         return func
 
     @testing.for_all_dtypes(no_bool=True, no_complex=True)
@@ -126,6 +129,7 @@ class TestReturnNone(FusionTestBase):
     def test_iadd_once(self, xp, dtype):
         def func(x):
             x += 2
+
         return func
 
     @testing.for_all_dtypes(no_bool=True, no_complex=True)
@@ -134,6 +138,7 @@ class TestReturnNone(FusionTestBase):
         def impl(x):
             x += x
             x += x
+
         return impl
 
 
@@ -162,6 +167,7 @@ class TestSpecialValues(FusionTestBase):
         def func(x):
             x[:] = cupy.nan
             return x
+
         return func
 
     @testing.for_float_dtypes()
@@ -170,6 +176,7 @@ class TestSpecialValues(FusionTestBase):
         def func(x):
             x[:] = cupy.inf
             return x
+
         return func
 
     @testing.for_float_dtypes()
@@ -178,9 +185,10 @@ class TestSpecialValues(FusionTestBase):
         def func(x):
             x[:] = -cupy.inf
             return x
+
         return func
 
-    @testing.with_requires('numpy>=2.0')
+    @testing.with_requires("numpy>=2.0")
     @testing.for_float_dtypes()
     @testing.numpy_cupy_allclose()
     def test_inf_where(self, xp, dtype):
@@ -193,7 +201,7 @@ class TestSpecialValues(FusionTestBase):
         mask = testing.shaped_random((N,), xp=xp, dtype=dtype) > 0.5
         return func(points, mask)
 
-    @testing.with_requires('numpy>=2.0')
+    @testing.with_requires("numpy>=2.0")
     @testing.for_float_dtypes()
     @testing.numpy_cupy_allclose()
     def test_nan_where(self, xp, dtype):
@@ -214,8 +222,8 @@ class TestFusionDecorator(unittest.TestCase):
             """Fuse without parentheses"""
             return x + x
 
-        assert func_wo_paren.__name__ == 'func_wo_paren'
-        assert func_wo_paren.__doc__ == 'Fuse without parentheses'
+        assert func_wo_paren.__name__ == "func_wo_paren"
+        assert func_wo_paren.__doc__ == "Fuse without parentheses"
 
     def test_with_paren(self):
         @cupy.fuse()
@@ -223,22 +231,21 @@ class TestFusionDecorator(unittest.TestCase):
             """Fuse with parentheses"""
             return x + x
 
-        assert func_w_paren.__name__ == 'func_w_paren'
-        assert func_w_paren.__doc__ == 'Fuse with parentheses'
+        assert func_w_paren.__name__ == "func_w_paren"
+        assert func_w_paren.__doc__ == "Fuse with parentheses"
 
 
 class TestFusionKernelName(unittest.TestCase):
 
     def check(self, xp, func, expected_name, is_elementwise):
-        a = xp.arange(0, 12, dtype='d').reshape(3, 4)
-        b = xp.arange(5, 17, dtype='d').reshape(3, 4)
-        c = xp.arange(13, 25, dtype='d').reshape(3, 4)
+        a = xp.arange(0, 12, dtype="d").reshape(3, 4)
+        b = xp.arange(5, 17, dtype="d").reshape(3, 4)
+        c = xp.arange(13, 25, dtype="d").reshape(3, 4)
 
         # Test kernel name (with mock)
         if xp is cupy:
             target = cupy._core._fusion_kernel.FusedKernel
-            target_full_name = '{}.{}'.format(
-                target.__module__, target.__name__)
+            target_full_name = "{}.{}".format(target.__module__, target.__name__)
 
             with mock.patch(target_full_name) as kernel:  # NOQA
                 func(a, b, c)
@@ -258,18 +265,18 @@ class TestFusionKernelName(unittest.TestCase):
 
             return func_a1(a, b, c)
 
-        return self.check(xp, func, 'func_a1', True)
+        return self.check(xp, func, "func_a1", True)
 
     @testing.numpy_cupy_allclose(atol=1e-5)
     def test_elementwise_with_name(self, xp):
         def func(a, b, c):
-            @cupy.fuse(kernel_name='abc')
+            @cupy.fuse(kernel_name="abc")
             def func_a1(x, y, z):
                 return (x + y) * z
 
             return func_a1(a, b, c)
 
-        return self.check(xp, func, 'abc', True)
+        return self.check(xp, func, "abc", True)
 
     @testing.numpy_cupy_allclose(atol=1e-5)
     def test_reduction_premap(self, xp):
@@ -280,7 +287,7 @@ class TestFusionKernelName(unittest.TestCase):
 
             return func_a1(a, b, c)
 
-        return self.check(xp, func, 'func_a1', False)
+        return self.check(xp, func, "func_a1", False)
 
     @testing.numpy_cupy_allclose(atol=1e-5)
     def test_reduction_postmap(self, xp):
@@ -291,7 +298,7 @@ class TestFusionKernelName(unittest.TestCase):
 
             return func_a1(a)
 
-        return self.check(xp, func, 'func_a1', False)
+        return self.check(xp, func, "func_a1", False)
 
     @testing.numpy_cupy_allclose(atol=1e-5)
     def test_reduction_01(self, xp):
@@ -302,18 +309,18 @@ class TestFusionKernelName(unittest.TestCase):
 
             return func_a1(a, b, c)
 
-        return self.check(xp, func, 'func_a1', False)
+        return self.check(xp, func, "func_a1", False)
 
     @testing.numpy_cupy_allclose(atol=1e-5)
     def test_reduction_with_name(self, xp):
         def func(a, b, c):
-            @cupy.fuse(kernel_name='abc')
+            @cupy.fuse(kernel_name="abc")
             def func_a1(x, y, z):
                 return xp.sum((x + y) * z)
 
             return func_a1(a, b, c)
 
-        return self.check(xp, func, 'abc', False)
+        return self.check(xp, func, "abc", False)
 
 
 class TestFusionComposition(unittest.TestCase):
